Replaced EQTemplate with EQRemover
Allows automatic reading of catalog from FDSN server, saves to local file
for future runs.  Interface changed to be more modular.  No longer saves
a Trace with zeros and ones but rather a list of time periods in which the
values of another Trace should be set to zero using the zero() method.